[refactor]:HomeTab컴포넌트 수정하기
